consolidate chat system into deep ChatService module
